+++ test/Servant/MatrixParam/ClientSpec.hs
@@ -55,16 +55,27 @@ spec = do
 
 
 #if MIN_VERSION_servant_client(0,9,0)
-    it "generates paths that servant-server understands" $ do
+    it "generates paths that servant-server understands: cliA" $ do
       testWithApplication (return $ serve api server) $ \port -> do
-        let res = (,,) <$> cliA (Just "There is a there there")
-                       <*> cliB (Just 1) (Just 2)
-                       <*> cliC 5 (Just 1) True
-
+        let res = cliA (Just "There is a there there")
             url = BaseUrl Http "localhost" port ""
         mgr' <- newManager defaultManagerSettings
-        runClientM res (ClientEnv mgr' url)
-          `shouldReturn` Right ("There is a there there", "3", "6")
+        runClientM res (ClientEnv mgr' url Nothing)
+          `shouldReturn` Right "There is a there there"
+    it "generates paths that servant-server understands: cliB" $ do
+      testWithApplication (return $ serve api server) $ \port -> do
+        let res = cliB (Just 1) (Just 2)
+            url = BaseUrl Http "localhost" port ""
+        mgr' <- newManager defaultManagerSettings
+        runClientM res (ClientEnv mgr' url Nothing)
+          `shouldReturn` Right "3"
+    it "generates paths that servant-server understands: cliC" $ do
+      testWithApplication (return $ serve api server) $ \port -> do
+        let res = cliC 5 (Just 1) True
+            url = BaseUrl Http "localhost" port ""
+        mgr' <- newManager defaultManagerSettings
+        runClientM res (ClientEnv mgr' url Nothing)
+          `shouldReturn` Right "6"
 #endif
 ------------------------------------------------------------------------------
 -- API
@@ -104,7 +115,7 @@ hasRequestPath :: (Manager -> BaseUrl -> ClientM any) -> ByteString -> IO ()
 hasRequestPath c expectedPath = do
   let anyBurl = BaseUrl Http "localhost" 6660 ""
 #if MIN_VERSION_servant_client(0,9,0)
-  _ <- void (runClientM c $ ClientEnv mgr anyBurl)
+  _ <- void (runClientM c $ ClientEnv mgr anyBurl Nothing)
     `catch` \(_ :: SomeException) -> return ()
 #else
   _ <- void (runExceptT $ c mgr anyBurl)