+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Revision History for servant-matrix-param
+
+# 0.4.0 -- 2018-5-14
+- upgrade to ghc 8.2.2, lts-11.7
+
+# 0.3.6 -- 2018-04-26
+- tests building
+
+# 0.3.5 -- 2018-04-15
+- Fix Client and server with new requirements for new servant versions (>12)
+
 # 0.3.4 -- 2018-01-04
 
 - Relax dependency version for servant-aeson-specs.